FixedHuman In the Loop Middleware throws error
